+++ handbook/src/pages/index.js
@@ -144,7 +144,7 @@ function Gitee() {
             className={"furion-log-jiao" + (isDarkTheme ? " dark" : "")}
           ></div>
           <div className="furion-log-number">
-            <div style={{ color: "#3fbbfe" }}>1300</div>
+            <div style={{ color: "#3fbbfe" }}>1350</div>
             <span className={isDarkTheme ? " dark" : ""}>Forks</span>
           </div>
         </div>
@@ -156,7 +156,7 @@ function Gitee() {
             className={"furion-log-jiao" + (isDarkTheme ? " dark" : "")}
           ></div>
           <div className="furion-log-number">
-            <div style={{ color: "#1fd898" }}>150,249</div>
+            <div style={{ color: "#1fd898" }}>150,413</div>
             <span className={isDarkTheme ? " dark" : ""}>Downloads</span>
           </div>
         </div>